Update main.py
Complete the generation process.
Parameters 'beam_size', 'max_gen_len' and so on can be used to control the generation method and length. These can be found at dict data 'generation_configs'.
When we do test for generation, half of the SMILES will be the input sequence and the other half will be predicted, processed, concatenated with the original input (the first half of the SMILES) and decoded from tensors to SMILES. Through parameters 'max_gen_len', we can test generation on different lengths, with processed and caoncatenated SMILES as output in our terminal.
Last, a visualization function is added to contrast the tokens generated.